# adapters/api/server/server.gen.go
//go:build go1.22

// Package server provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version 2.5.1 DO NOT EDIT.
package server

import (
	"fmt"
	"net/http"

	"github.com/oapi-codegen/runtime"
)

// BoundingBox defines model for BoundingBox.
type BoundingBox struct {
	// Height height of the bounding box
	Height float32 `json:"height"`

	// Id ID of the bounding box
	Id string `json:"id"`

	// Label label
	Label string `json:"label"`

	// Width width of the bounding box
	Width float32 `json:"width"`

	// Xc x coordinate of the center point
	Xc float32 `json:"xc"`

	// Yc y coordinate of the center point
	Yc float32 `json:"yc"`
}

// Collection defines model for Collection.
type Collection struct {
	// Description Description of the collection
	Description *string `json:"description,omitempty"`

	// Name Name of the collection
	Name *string `json:"name,omitempty"`
}

// Error defines model for Error.
type Error struct {
	// Code Error code
	Code int32 `json:"code"`

	// Message Error message
	Message string `json:"message"`
}

// Image defines model for Image.
type Image struct {
	BoundingBoxes *[]BoundingBox `json:"bounding_boxes,omitempty"`

	// Collection name of collection in which the image belongs
	Collection string `json:"collection"`

	// Id ID of the image
	Id     string    `json:"id"`
	Labels *[]string `json:"labels,omitempty"`
}

// ImageIngestionResponse defines model for ImageIngestionResponse.
type ImageIngestionResponse struct {
	// Id ID of ingested image
	Id *string `json:"id,omitempty"`
}

// Label defines model for Label.
type Label struct {
	// Description Description of the label
	Description *string `json:"description,omitempty"`

	// Name Name of the label
	Name *string `json:"name,omitempty"`
}

// ListCollectionsResponse defines model for ListCollectionsResponse.
type ListCollectionsResponse struct {
	Data       *[]Collection `json:"data,omitempty"`
	Pagination Pagination    `json:"pagination"`
}

// ListImagesResponse defines model for ListImagesResponse.
type ListImagesResponse struct {
	Images     []Image    `json:"images"`
	Pagination Pagination `json:"pagination"`
}

// ListLabelsResponse defines model for ListLabelsResponse.
type ListLabelsResponse struct {
	Data       *[]Label   `json:"data,omitempty"`
	Pagination Pagination `json:"pagination"`
}

// NewBoundingBox defines model for NewBoundingBox.
type NewBoundingBox struct {
	// Height height of the bounding box
	Height float32 `json:"height"`

	// Label label
	Label string `json:"label"`

	// Width width of the bounding box
	Width float32 `json:"width"`

	// Xc x coordinate of the center point
	Xc float32 `json:"xc"`

	// Yc y coordinate of the center point
	Yc float32 `json:"yc"`
}

// NewCollection defines model for NewCollection.
type NewCollection struct {
	// Description Description of the collection
	Description *string `json:"description,omitempty"`

	// Name Name of the collection
	Name string `json:"name"`
}

// NewImage defines model for NewImage.
type NewImage struct {
	BoundingBoxes *[]NewBoundingBox `json:"bounding_boxes,omitempty"`

	// Collection name of collection in which to add the image
	Collection string `json:"collection"`

	// Data base64 encoded image raw bytes
	Data   string    `json:"data"`
	Labels *[]string `json:"labels,omitempty"`
}

// NewLabel defines model for NewLabel.
type NewLabel struct {
	// Description Description of the label
	Description *string `json:"description,omitempty"`

	// Name Name of the label
	Name string `json:"name"`
}

// Pagination defines model for Pagination.
type Pagination struct {
	// Page current page number
	Page int64 `json:"page"`

	// PageSize maximum number of items per page
	PageSize int `json:"page_size"`

	// TotalItems total number of items
	TotalItems int64 `json:"total_items"`

	// TotalPages total number of pages
	TotalPages int64 `json:"total_pages"`
}

// UpdateCollection defines model for UpdateCollection.
type UpdateCollection struct {
	// Description New description of the collection
	Description string `json:"description"`

	// Name New name of the collection
	Name string `json:"name"`
}

// ListCollectionsParams defines parameters for ListCollections.
type ListCollectionsParams struct {
	// Page page number
	Page *int64 `form:"page,omitempty" json:"page,omitempty"`

	// PageSize maximum number of collections to return
	PageSize *int `form:"page_size,omitempty" json:"page_size,omitempty"`
}

// ListImagesParams defines parameters for ListImages.
type ListImagesParams struct {
	// Page page number
	Page *int64 `form:"page,omitempty" json:"page,omitempty"`

	// PageSize maximum number of collections to return
	PageSize *int `form:"page_size,omitempty" json:"page_size,omitempty"`

	// Collection name of collection
	Collection *string `form:"collection,omitempty" json:"collection,omitempty"`
}

// ListLabelsParams defines parameters for ListLabels.
type ListLabelsParams struct {
	// Page page number
	Page *int64 `form:"page,omitempty" json:"page,omitempty"`

	// PageSize maximum number of labels to return
	PageSize *int `form:"page_size,omitempty" json:"page_size,omitempty"`
}

// CreateCollectionJSONRequestBody defines body for CreateCollection for application/json ContentType.
type CreateCollectionJSONRequestBody = NewCollection

// UpdateCollectionByNameJSONRequestBody defines body for UpdateCollectionByName for application/json ContentType.
type UpdateCollectionByNameJSONRequestBody = UpdateCollection

// IngestImageJSONRequestBody defines body for IngestImage for application/json ContentType.
type IngestImageJSONRequestBody = NewImage

// CreateLabelJSONRequestBody defines body for CreateLabel for application/json ContentType.
type CreateLabelJSONRequestBody = NewLabel

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// List collections
	// (GET /collections)
	ListCollections(w http.ResponseWriter, r *http.Request, params ListCollectionsParams)
	// Create a new collection
	// (POST /collections)
	CreateCollection(w http.ResponseWriter, r *http.Request)
	// Delete a collection by name
	// (DELETE /collections/{name})
	DeleteCollectionByName(w http.ResponseWriter, r *http.Request, name string)
	// Find a collection by name
	// (GET /collections/{name})
	FindCollectionByName(w http.ResponseWriter, r *http.Request, name string)
	// Update a collection
	// (PUT /collections/{name})
	UpdateCollectionByName(w http.ResponseWriter, r *http.Request, name string)
	// List images
	// (GET /images)
	ListImages(w http.ResponseWriter, r *http.Request, params ListImagesParams)
	// Ingest a new image
	// (POST /images)
	IngestImage(w http.ResponseWriter, r *http.Request)
	// Read image meta-data
	// (GET /images/{collection_name}/{image_id})
	ReadImage(w http.ResponseWriter, r *http.Request, collectionName string, imageId string)
	// List labels
	// (GET /labels)
	ListLabels(w http.ResponseWriter, r *http.Request, params ListLabelsParams)
	// Create a new label
	// (POST /labels)
	CreateLabel(w http.ResponseWriter, r *http.Request)
	// Delete a label by name
	// (DELETE /labels/{name})
	DeleteLabelByName(w http.ResponseWriter, r *http.Request, name string)
	// Find a label by name
	// (GET /labels/{name})
	FindLabelByName(w http.ResponseWriter, r *http.Request, name string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// ListCollections operation middleware
func (siw *ServerInterfaceWrapper) ListCollections(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ListCollectionsParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", r.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page", Err: err})
		return
	}

	// ------------- Optional query parameter "page_size" -------------

	err = runtime.BindQueryParameter("form", true, false, "page_size", r.URL.Query(), &params.PageSize)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page_size", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListCollections(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateCollection operation middleware
func (siw *ServerInterfaceWrapper) CreateCollection(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateCollection(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DeleteCollectionByName operation middleware
func (siw *ServerInterfaceWrapper) DeleteCollectionByName(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "name" -------------
	var name string

	err = runtime.BindStyledParameterWithOptions("simple", "name", r.PathValue("name"), &name, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "name", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteCollectionByName(w, r, name)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// FindCollectionByName operation middleware
func (siw *ServerInterfaceWrapper) FindCollectionByName(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "name" -------------
	var name string

	err = runtime.BindStyledParameterWithOptions("simple", "name", r.PathValue("name"), &name, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "name", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.FindCollectionByName(w, r, name)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UpdateCollectionByName operation middleware
func (siw *ServerInterfaceWrapper) UpdateCollectionByName(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "name" -------------
	var name string

	err = runtime.BindStyledParameterWithOptions("simple", "name", r.PathValue("name"), &name, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "name", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UpdateCollectionByName(w, r, name)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListImages operation middleware
func (siw *ServerInterfaceWrapper) ListImages(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ListImagesParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", r.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page", Err: err})
		return
	}

	// ------------- Optional query parameter "page_size" -------------

	err = runtime.BindQueryParameter("form", true, false, "page_size", r.URL.Query(), &params.PageSize)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page_size", Err: err})
		return
	}

	// ------------- Optional query parameter "collection" -------------

	err = runtime.BindQueryParameter("form", true, false, "collection", r.URL.Query(), &params.Collection)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "collection", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListImages(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// IngestImage operation middleware
func (siw *ServerInterfaceWrapper) IngestImage(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.IngestImage(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ReadImage operation middleware
func (siw *ServerInterfaceWrapper) ReadImage(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "collection_name" -------------
	var collectionName string

	err = runtime.BindStyledParameterWithOptions("simple", "collection_name", r.PathValue("collection_name"), &collectionName, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "collection_name", Err: err})
		return
	}

	// ------------- Path parameter "image_id" -------------
	var imageId string

	err = runtime.BindStyledParameterWithOptions("simple", "image_id", r.PathValue("image_id"), &imageId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "image_id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ReadImage(w, r, collectionName, imageId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListLabels operation middleware
func (siw *ServerInterfaceWrapper) ListLabels(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ListLabelsParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", r.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page", Err: err})
		return
	}

	// ------------- Optional query parameter "page_size" -------------

	err = runtime.BindQueryParameter("form", true, false, "page_size", r.URL.Query(), &params.PageSize)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page_size", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListLabels(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateLabel operation middleware
func (siw *ServerInterfaceWrapper) CreateLabel(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateLabel(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DeleteLabelByName operation middleware
func (siw *ServerInterfaceWrapper) DeleteLabelByName(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "name" -------------
	var name string

	err = runtime.BindStyledParameterWithOptions("simple", "name", r.PathValue("name"), &name, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "name", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteLabelByName(w, r, name)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// FindLabelByName operation middleware
func (siw *ServerInterfaceWrapper) FindLabelByName(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "name" -------------
	var name string

	err = runtime.BindStyledParameterWithOptions("simple", "name", r.PathValue("name"), &name, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "name", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.FindLabelByName(w, r, name)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of http.ServeMux.
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	ServeHTTP(w http.ResponseWriter, r *http.Request)
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc("GET "+options.BaseURL+"/collections", wrapper.ListCollections)
	m.HandleFunc("POST "+options.BaseURL+"/collections", wrapper.CreateCollection)
	m.HandleFunc("DELETE "+options.BaseURL+"/collections/{name}", wrapper.DeleteCollectionByName)
	m.HandleFunc("GET "+options.BaseURL+"/collections/{name}", wrapper.FindCollectionByName)
	m.HandleFunc("PUT "+options.BaseURL+"/collections/{name}", wrapper.UpdateCollectionByName)
	m.HandleFunc("GET "+options.BaseURL+"/images", wrapper.ListImages)
	m.HandleFunc("POST "+options.BaseURL+"/images", wrapper.IngestImage)
	m.HandleFunc("GET "+options.BaseURL+"/images/{collection_name}/{image_id}", wrapper.ReadImage)
	m.HandleFunc("GET "+options.BaseURL+"/labels", wrapper.ListLabels)
	m.HandleFunc("POST "+options.BaseURL+"/labels", wrapper.CreateLabel)
	m.HandleFunc("DELETE "+options.BaseURL+"/labels/{name}", wrapper.DeleteLabelByName)
	m.HandleFunc("GET "+options.BaseURL+"/labels/{name}", wrapper.FindLabelByName)

	return m
}