# site/openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: go-image-annotator API
  description: Image Annotation Service
  termsOfService: https://swagger.io/terms/
  contact:
    name: Swagger API Team
    email: apiteam@swagger.io
    url: https://swagger.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /api
paths:
  /images/{collection_name}/{image_id}:
    get:
      summary: Read image meta-data
      description: Read meta-data of an image
      operationId: readImage
      tags: [Image]
      parameters:
        - name: collection_name
          in: path
          description: Name of collection
          required: true
          schema:
            type: string
        - name: image_id
          in: path
          description: ID of image
          required: true
          schema:
            type: string
      responses:
        '200':
          description: image meta-data response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /images:
    get:
      summary: List images
      description: Paginate images
      operationId: listImages
      tags: [Image]
      parameters:
        - name: page
          in: query
          description: page number
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: maximum number of collections to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: collection
          in: query
          description: name of collection
          required: false
          schema:
            type: string
      responses:
        '200':
          description: list image response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListImagesResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Ingest a new image
      description: Ingest an image into a collection. 
      operationId: ingestImage
      tags: [Image]
      requestBody:
        description: Ingestion payload
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewImage'
      responses:
        '200':
          description: image ingestion response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageIngestionResponse'
        default:
          description: unexpected error
          content:
            application/json:
            schema:
                $ref: '#/components/schemas/Error'
  /collections/{name}:
    get:
      summary: Find a collection by name
      description: Returns a collection based on its name
      operationId: findCollectionByName
      tags: [Collection]
      parameters:
        - name: name
          in: path
          description: Name of collection to fetch
          required: true
          schema:
            type: string
      responses:
        '200':
          description: collection response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a collection
      description: Update the name and description of a collection
      operationId: updateCollectionByName
      tags: [Collection]
      parameters:
        - name: name
          in: path
          required: true
          description: Current name of the collection
          schema:
            type: string
      requestBody:
        description: New values for the collection
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCollection'
      responses:
        '200':
          description: Collection updated successfully
        '404':
          description: Collection not found
        '400':
          description: Invalid input
    delete:
      summary: Delete a collection by name
      description: Delete a collection based on its name
      operationId: deleteCollectionByName
      tags: [Collection]
      parameters:
        - name: name
          in: path
          description: Name of collection to delete
          required: true
          schema:
            type: string
      responses:
        '204':
          description: collection deleted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /collections:
    get:
      summary: List collections
      description: Paginate collections
      operationId: listCollections
      tags: [Collection]
      parameters:
        - name: page
          in: query
          description: page number
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: maximum number of collections to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: list collection response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListCollectionsResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new collection
      description: Creates a new collection. Duplicate names are not allowed
      operationId: createCollection
      tags: [Collection]
      requestBody:
        description: Collection to add
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCollection'
      responses:
        '200':
          description: collection response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /labels/{name}:
    get:
      summary: Find a label by name
      description: Returns a label based on its name
      operationId: findLabelByName
      tags: [Label]
      parameters:
        - name: name
          in: path
          description: Name of label to fetch
          required: true
          schema:
            type: string
      responses:
        '200':
          description: label response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Label'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a label by name
      description: Delete a label based on its name
      operationId: deleteLabelByName
      tags: [Label]
      parameters:
        - name: name
          in: path
          description: Name of label to delete
          required: true
          schema:
            type: string
      responses:
        '204':
          description: label deleted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /labels:
    get:
      summary: List labels
      description: Returns all pets from the system that the user has access to
      operationId: listLabels
      tags: [Label]
      parameters:
        - name: page
          in: query
          description: page number
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: maximum number of labels to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: list label response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListLabelsResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new label
      description: Creates a new label. Duplicate names are not allowed
      operationId: createLabel
      tags: [Label]
      requestBody:
        description: Label to add
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewLabel'
      responses:
        '200':
          description: label response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Label'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    Collection:
      properties:
        name:
          type: string
          description: Name of the collection
        description:
          type: string
          description: Description of the collection
    UpdateCollection:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          description: New name of the collection
        description:
          type: string
          description: New description of the collection
    ListCollectionsResponse:
      type: object
      required:
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Collection'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ListImagesResponse:
      type: object
      required:
        - images
        - pagination
      properties:
        images:
          type: array
          items:
            $ref: '#/components/schemas/Image'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Image:
      required:
        - id
        - collection
      properties:
        id:
          type: string
          description: ID of the image
        collection:
          type: string
          description: name of collection in which the image belongs
        labels:
          type: array
          items:
            type: string
        bounding_boxes:
          type: array
          items:
            $ref: '#/components/schemas/BoundingBox'
    NewImage:
      required:
        - data
        - collection
      properties:
        data:
          type: string
          description: base64 encoded image raw bytes
        collection:
          type: string
          description: name of collection in which to add the image
        labels:
          type: array
          items:
            type: string
        bounding_boxes:
          type: array
          items:
            $ref: '#/components/schemas/NewBoundingBox'
...
    ImageIngestionResponse:
      properties:
        id:
          type: string
          description: ID of ingested image
    NewCollection:
      required:
        - name
      properties:
        name:
          type: string
          description: Name of the collection
        description:
          type: string
          description: Description of the collection
    Label:
      properties:
        name:
          type: string
          description: Name of the label
        description:
          type: string
          description: Description of the label
    ListLabelsResponse:
      type: object
      required:
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        pagination:
          $ref: '#/components/schemas/Pagination'
    NewLabel:
      required:
        - name
      properties:
        name:
          type: string
          description: Name of the label
        description:
          type: string
          description: Description of the label
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Error code
        message:
          type: string
          description: Error message